(CTH-297) Use transaction_id to refer to action tasks
Using the transaction_id of non-blocking requests to refer to remote
tasks, instead of using an ad-hoc job_id entry.

We now simplify the provisional response schema; only the transaction_id
of the requested action is returned. The message itself will indicate
that the action has been successfully started. In case of failure when
starting a task, an RPC error message should be sent.