(maint) Update PXP non-blocking response schema v1
Here we update the schema of the PXP data content in the title to the
actual schema that was implemented originally (without the "job_id"
entry).

Note that we decided to use the transaction_id as the reference to
previously requested non-blocking actions; at that point, the job_id
became superflous. And it was not implemented in the first PXP
implementation.