(maint) Propose removing inventory version
Inventory version was intended to ensure that inventory updates would
not be missed or arive out-of-order. However, they didn't promise any
consistency when connection between client and broker is lost.

TCP guarantees in-order-delivery. That should be sufficient to ensure
that - over a single TCP session - no inventory updates are missed or
arrive out-of-order. As such, inventory version seems redundant.